+++ macos-leo/app/Components/Types/FloatingDockType.ts
@@ -3,11 +3,18 @@ export interface DockItem {
   title: string;
   imageSrc: string;
   href: string;
+  content: string;
 }
 
+export interface FloatingDockDesktopProps {
+  items: DockItem[];
+  className?: string;
+}
 
-  
-  export interface FloatingDockDesktopProps {
-    items: DockItem[];
-    className?: string;
-  }
+export interface FloatingDockProps {
+  items: DockItem[];
+  className?: string;
+  onItemClick: (id: string) => void;
+  openFolders: string[];
+  minimizedFolders: string[];
+}