+++ macos-leo/app/page.tsx
@@ -1,11 +1,10 @@
 "use client";
 
-import { DockItems } from "./Components/Dock/DockItems";
-import { FloatingDock } from "./Components/Dock/FloatingDock";
 import MorngBg from "@/public/assets/background/wallpaper-day.jpg";
 import NightBg from "@/public/assets/background/wallpaper-night.jpg";
 import Navbar from "./Components/Navbar";
 import { UserThemeStore } from "./state/store";
+import MainScreen from "./Components/MainScreen/MainScreen";
 
 export default function Home() {
   const { setIsDark, isDark } = UserThemeStore();
@@ -22,13 +21,12 @@ export default function Home() {
       }}
     >
       <Navbar />
+      {/* FOLDER SPACE */}
       {/* fixed top Navbar */}
       {/* CENTER FOR DISPLAYING  */}
       {/* fixed bottom Dock Bar */}
-      <div className="fixed bottom-0 z-50 flex justify-center items-center w-full">
-        {/* <p>Hellooooo Guyss</p> */}
-        <FloatingDock items={DockItems} />
-      </div>
+      <MainScreen/>
+
     </div>
   );
 }